moved auth validation exception to its own namespace
